# specs/008-duckdock-2-foundation/contracts/openapi-v2.yaml
openapi: 3.1.0
...
info:
  title: DuckDock Agent Execution API
  version: 2.0.0-draft
  summary: Reporter ingestion and governed AgentRun read contracts for DuckDock 2.0.
  description: |
    Foundation contract for metadata-only Agent Harness execution reporting.
    Reporter writes are authenticated with scoped Reporter bearer tokens and are
    idempotent. Namespace and Runtime identity are derived from the credential;
    they are never accepted from the Reporter request body. Raw prompts,
    completions, tool arguments, tool results, retrieved content, and hidden
    chain-of-thought are outside this API contract.
servers:
  - url: /
...
paths:
...
  /api/v2/agent-runs:
    get:
      tags:
        - Agent Runs
      summary: List governed Agent runs
      operationId: listAgentRuns
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
        - $ref: '#/components/parameters/RuntimeInstanceFilter'
        - $ref: '#/components/parameters/DeploymentPublicIdFilter'
        - $ref: '#/components/parameters/RunStatusFilter'
        - $ref: '#/components/parameters/StartedAfter'
        - $ref: '#/components/parameters/StartedBefore'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of AgentRun summaries visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/agent-runs/{run_public_id}:
    get:
      tags:
        - Agent Runs
      summary: Get one governed Agent run
      operationId: getAgentRun
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RunPublicId'
      responses:
        '200':
          description: AgentRun detail with metadata-only telemetry and evidence links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/agent-runs/{run_public_id}/artifacts:
    get:
      tags:
        - Agent Run Artifacts
      summary: List immutable artifact metadata for an Agent run
      operationId: listAgentRunArtifacts
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RunPublicId'
        - $ref: '#/components/parameters/NamespaceId'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Artifact metadata visible in the requested Namespace.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRunArtifact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
        - Agent Run Artifacts
      summary: Register immutable artifact metadata for an Agent run
      description: |
        Registers checksum, size, schema and an internal object key/URI only.
        DuckDock does not download or parse artifact content in this operation.
      operationId: registerAgentRunArtifact
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RunPublicId'
        - $ref: '#/components/parameters/NamespaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunArtifactCreate'
      responses:
        '201':
          description: Artifact metadata registered or idempotently replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunArtifact'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/telemetry-sinks:
    get:
      tags:
        - Telemetry Sinks
      summary: List provider-neutral telemetry sinks
      operationId: listTelemetrySinks
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
        - name: include_disabled
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Telemetry sinks visible in the requested Namespace.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetrySink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
        - Telemetry Sinks
      summary: Create a provider-neutral telemetry sink
      operationId: createTelemetrySink
      security:
        - UserBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetrySinkCreate'
      responses:
        '201':
          description: Telemetry sink created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetrySink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/telemetry-sinks/{public_id}:
    get:
      tags:
        - Telemetry Sinks
      summary: Get one telemetry sink
      operationId: getTelemetrySink
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TelemetrySinkPublicId'
      responses:
        '200':
          description: Telemetry sink detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetrySink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Telemetry Sinks
      summary: Update mutable telemetry sink configuration
      operationId: updateTelemetrySink
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TelemetrySinkPublicId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetrySinkUpdate'
      responses:
        '200':
          description: Telemetry sink updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetrySink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v2/telemetry-sinks/{public_id}/disable:
    post:
      tags:
        - Telemetry Sinks
      summary: Disable a telemetry sink
      operationId: disableTelemetrySink
      security:
        - UserBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TelemetrySinkPublicId'
      responses:
        '200':
          description: Telemetry sink disabled or already disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetrySink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ReporterBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: DuckDock Reporter Token
      description: |
        Scoped, rotatable Reporter credential bound to a Namespace and Runtime.
        The execution API requires the `execution.write` scope. The credential
        cannot be used as a user JWT or to read arbitrary AgentRuns.
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RunPublicId:
      name: run_public_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
    TelemetrySinkPublicId:
      name: public_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^tsk_[0-9a-f]{32}$'
...
    NamespaceId:
      name: namespace_id
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
    RuntimeInstanceFilter:
      name: runtime_instance_id
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    DeploymentPublicIdFilter:
      name: deployment_public_id
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    RunStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/RunStatus'
    StartedAfter:
      name: started_after
      in: query
      required: true
      description: Inclusive UTC lower bound for a query window of at most 31 days.
      schema:
        type: string
        format: date-time
    StartedBefore:
      name: started_before
      in: query
      required: true
      description: Exclusive UTC upper bound for a query window of at most 31 days.
      schema:
        type: string
        format: date-time
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 2000
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or incorrectly scoped bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource does not exist or is not visible in the caller scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Invalid lifecycle transition or Idempotency-Key payload conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Request violates the v2 lifecycle, correlation, or metadata-only contract.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    TelemetryEnvelopeV1:
      $ref: ./telemetry-envelope-v1.schema.json
...
    MetadataOnlyContentPolicy:
      type: string
      const: metadata_only
      description: |
        Only identifiers, timing, counts, status, model/tool names, token/cost
        summaries, and governed asset correlations may be submitted.
    MetadataOnlyCaptureMode:
      type: string
      const: metadata_only
      description: |
        The Reporter control envelope carries only allowlisted identifiers,
        lifecycle timestamps, counters, classified errors, and trace references.
    MetadataOnlyTelemetryEnvelope:
      allOf:
        - $ref: '#/components/schemas/TelemetryEnvelopeV1'
        - type: object
          required:
            - content_policy
          properties:
            content_policy:
              $ref: '#/components/schemas/MetadataOnlyContentPolicy'
...
    RunStatus:
      type: string
      enum:
        - STARTED
        - SUCCEEDED
        - FAILED
        - CANCELLED
        - TIMED_OUT
    Sensitivity:
      type: string
      enum:
        - PUBLIC
        - INTERNAL
        - CONFIDENTIAL
        - RESTRICTED
    AgentRunArtifactKind:
      type: string
      enum:
        - TRAJECTORY
        - EVALUATION
        - LOG_BUNDLE
        - OTHER
    AgentRunArtifactCompleteness:
      type: string
      enum:
        - COMPLETE
        - PARTIAL
        - UNKNOWN
    AgentRunArtifactCreate:
      type: object
      additionalProperties: false
      required:
        - kind
        - schema_name
        - schema_version
        - object_uri
        - sha256
        - size_bytes
      properties:
        kind:
          $ref: '#/components/schemas/AgentRunArtifactKind'
        schema_name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
        schema_version:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
        object_uri:
          type: string
          minLength: 1
          maxLength: 512
          description: Relative internal object key or URI in the configured s3 bucket.
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        size_bytes:
          type: integer
          minimum: 0
        sensitivity:
          $ref: '#/components/schemas/Sensitivity'
        redaction_policy_version:
          type: string
          minLength: 1
          maxLength: 64
        completeness:
          $ref: '#/components/schemas/AgentRunArtifactCompleteness'
    AgentRunArtifact:
      type: object
      additionalProperties: false
      required:
        - public_id
        - namespace_id
        - run_public_id
        - kind
        - schema_name
        - schema_version
        - object_uri
        - sha256
        - size_bytes
        - sensitivity
        - completeness
        - created_at
      properties:
        public_id:
          type: string
          pattern: '^art_[0-9a-f]{32}$'
        namespace_id:
          type: integer
          minimum: 1
        run_public_id:
          type: string
          minLength: 1
        kind:
          $ref: '#/components/schemas/AgentRunArtifactKind'
        schema_name:
          type: string
          minLength: 1
          maxLength: 100
        schema_version:
          type: string
          minLength: 1
          maxLength: 50
        object_uri:
          type: string
          minLength: 1
          maxLength: 512
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        size_bytes:
          type: integer
          minimum: 0
        sensitivity:
          $ref: '#/components/schemas/Sensitivity'
        redaction_policy_version:
          type:
            - string
            - 'null'
          maxLength: 64
        completeness:
          $ref: '#/components/schemas/AgentRunArtifactCompleteness'
        created_at:
          type: string
          format: date-time
    TelemetrySinkStatus:
      type: string
      enum:
        - ACTIVE
        - DISABLED
        - ERROR
    TelemetrySinkConfig:
      type: object
      additionalProperties: false
      properties:
        protocol:
          type: string
          enum:
            - native
            - otlp_http
            - otlp_grpc
            - custom
          default: custom
        environment:
          type: string
          minLength: 1
          maxLength: 64
        region:
          type: string
          minLength: 1
          maxLength: 64
        timeout_ms:
          type: integer
          minimum: 100
          maximum: 60000
          default: 15000
        verify_tls:
          type: boolean
          default: true
    TelemetrySinkCreate:
      type: object
      additionalProperties: false
      required:
        - namespace_id
        - provider
        - name
        - endpoint
        - credential_ref
      properties:
        namespace_id:
          type: integer
          minimum: 1
        provider:
          type: string
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
        name:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
        endpoint:
          type: string
          format: uri
          maxLength: 512
          description: SSRF-validated http/https endpoint without credentials or query values.
        project_ref:
          type: string
          minLength: 1
          maxLength: 255
        credential_ref:
          type: string
          maxLength: 255
          pattern: '^(db:[1-9][0-9]*|(?:secret|vault)://[A-Za-z0-9][A-Za-z0-9._/-]*)$'
        config:
          $ref: '#/components/schemas/TelemetrySinkConfig'
    TelemetrySinkUpdate:
      type: object
      additionalProperties: false
      properties:
        endpoint:
          type: string
          format: uri
          maxLength: 512
        project_ref:
          type: string
          minLength: 1
          maxLength: 255
        credential_ref:
          type: string
          maxLength: 255
          pattern: '^(db:[1-9][0-9]*|(?:secret|vault)://[A-Za-z0-9][A-Za-z0-9._/-]*)$'
        config:
          $ref: '#/components/schemas/TelemetrySinkConfig'
    TelemetrySink:
      type: object
      additionalProperties: false
      required:
        - public_id
        - namespace_id
        - provider
        - name
        - endpoint
        - credential_ref
        - status
        - config_json
        - created_at
        - updated_at
      properties:
        public_id:
          type: string
          pattern: '^tsk_[0-9a-f]{32}$'
        namespace_id:
          type: integer
          minimum: 1
        provider:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 128
        endpoint:
          type: string
          format: uri
          maxLength: 512
        project_ref:
          type:
            - string
            - 'null'
          maxLength: 255
        credential_ref:
          type: string
          maxLength: 255
        status:
          $ref: '#/components/schemas/TelemetrySinkStatus'
        config_json:
          $ref: '#/components/schemas/TelemetrySinkConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TrustLevel:
      type: string
      enum:
        - CHANNEL_AUTHENTICATED
        - PRODUCER_ATTESTED
        - UNVERIFIED
    TrustSource:
      type: string
      enum:
        - REPORTER
        - COLLECTOR
        - IMPORT
        - ADMIN
    TraceId:
      type: string
      pattern: '^[0-9a-f]{32}$'
    SpanId:
      type: string
      pattern: '^[0-9a-f]{16}$'
    MetadataAttributes:
      type: object
      maxProperties: 32
      propertyNames:
        type: string
        enum:
          - deployment.environment
          - deployment.region
          - harness.name
          - harness.version
          - service.name
          - service.version
          - operation.name
          - error.category
          - agent.mode
      additionalProperties:
        oneOf:
          - type: string
            maxLength: 2000
          - type: number
          - type: boolean
          - type: array
            maxItems: 100
            items:
              oneOf:
                - type: string
                  maxLength: 500
                - type: number
                - type: boolean
...
    AgentRunSummary:
      type: object
      additionalProperties: false
      required:
        - run_public_id
        - namespace_id
        - runtime_instance_id
        - external_run_id
        - status
        - trust_level
        - trust_source
        - source_schema
        - source_schema_version
        - normalizer_version
        - attempt
        - started_at
        - content_capture_mode
      properties:
        run_public_id:
          type: string
          minLength: 1
        external_run_id:
          type: string
          minLength: 1
        namespace_id:
          type: string
          minLength: 1
        runtime_instance_id:
          type: string
          minLength: 1
        session_public_id:
          type: string
          minLength: 1
        deployment_public_id:
          type: string
          minLength: 1
        work_trace_id:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/RunStatus'
        trust_level:
          $ref: '#/components/schemas/TrustLevel'
        trust_source:
          $ref: '#/components/schemas/TrustSource'
        source_schema:
          type: string
          minLength: 1
        source_schema_version:
          type: string
          minLength: 1
        normalizer_version:
          type: string
          minLength: 1
        otel_trace_id:
          $ref: '#/components/schemas/TraceId'
        root_span_id:
          $ref: '#/components/schemas/SpanId'
        attempt:
          type: integer
          minimum: 1
        started_at:
          type: string
          format: date-time
        ended_at:
          type:
            - string
            - 'null'
          format: date-time
        duration_ms:
          type: integer
          minimum: 0
        error_count:
          type: integer
          minimum: 0
        step_count:
          type: integer
          minimum: 0
        model_call_count:
          type: integer
          minimum: 0
        tool_call_count:
          type: integer
          minimum: 0
        input_token_count:
          type: integer
          minimum: 0
        output_token_count:
          type: integer
          minimum: 0
        error_type:
          type: string
          minLength: 1
          maxLength: 100
        content_capture_mode:
          $ref: '#/components/schemas/MetadataOnlyCaptureMode'
        metadata:
          $ref: '#/components/schemas/MetadataAttributes'
    AgentRun:
      type: object
      additionalProperties: false
      required:
        - run_public_id
        - namespace_id
        - runtime_instance_id
        - external_run_id
        - status
        - trust_level
        - trust_source
        - source_schema
        - source_schema_version
        - normalizer_version
        - attempt
        - started_at
        - content_capture_mode
      properties:
        run_public_id:
          type: string
          minLength: 1
        external_run_id:
          type: string
          minLength: 1
        namespace_id:
          type: string
          minLength: 1
        runtime_instance_id:
          type: string
          minLength: 1
        session_public_id:
          type: string
          minLength: 1
        deployment_public_id:
          type: string
          minLength: 1
        work_trace_id:
          type: string
          minLength: 1
        status:
          $ref: '#/components/schemas/RunStatus'
        trust_level:
          $ref: '#/components/schemas/TrustLevel'
        trust_source:
          $ref: '#/components/schemas/TrustSource'
        source_schema:
          type: string
          minLength: 1
        source_schema_version:
          type: string
          minLength: 1
        normalizer_version:
          type: string
          minLength: 1
        otel_trace_id:
          $ref: '#/components/schemas/TraceId'
        root_span_id:
          $ref: '#/components/schemas/SpanId'
        attempt:
          type: integer
          minimum: 1
        started_at:
          type: string
          format: date-time
        ended_at:
          type:
            - string
            - 'null'
          format: date-time
        duration_ms:
          type: integer
          minimum: 0
        step_count:
          type: integer
          minimum: 0
        model_call_count:
          type: integer
          minimum: 0
        tool_call_count:
          type: integer
          minimum: 0
        input_token_count:
          type: integer
          minimum: 0
        output_token_count:
          type: integer
          minimum: 0
        error_type:
          type: string
          minLength: 1
          maxLength: 100
        content_capture_mode:
          $ref: '#/components/schemas/MetadataOnlyCaptureMode'
        metadata:
          $ref: '#/components/schemas/MetadataAttributes'
        telemetry_summary:
          $ref: '#/components/schemas/MetadataOnlyTelemetryEnvelope'
    AgentRunList:
      type: object
      additionalProperties: false
      required:
        - items
        - next_cursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentRunSummary'
        next_cursor:
          type:
            - string
            - 'null'
    Problem:
      type: object
      additionalProperties: false
      required:
        - status
        - code
        - detail
        - request_id
        - retryable
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          minLength: 1
          maxLength: 100
        detail:
          type: string
          minLength: 1
          maxLength: 4000
        request_id:
          type: string
          minLength: 1
        retryable:
          type: boolean
        details:
          type: object
          maxProperties: 6
          propertyNames:
            type: string
            enum:
              - operation
              - field
              - limit
              - retry_after_seconds
              - dependency
              - conflict_type
          additionalProperties:
            oneOf:
              - type: string
                maxLength: 500
              - type: number
              - type: boolean
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - location
              - message
            properties:
              location:
                type: string
                minLength: 1
              message:
                type: string
                minLength: 1